# backend/api/openapi-bundled.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Multi Outlet POS Kopi Kopi & Inventory API",
    "version": "1.0.0",
    "description": "API Specification for Multi Outlet POS Kopi Popi & Inventory System."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local Development Server"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
    "/auth/register": {
      "post": {
        "summary": "Register Customer",
        "tags": [
          "Auth"
        ],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Gilang"
                  },
                  "email": {
                    "type": "string",
                    "example": "gilang@gmail.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "rahasia123"
                  },
                  "confirm_password": {
                    "type": "string",
                    "example": "rahasia123"
                  },
                  "phone": {
                    "type": "string",
                    "example": "08123456789"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "example": "uuid"
                        },
                        "name": {
                          "type": "string",
                          "example": "Gilang"
                        },
                        "email": {
                          "type": "string",
                          "example": "gilang@gmail.com"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Login",
        "tags": [
          "Auth"
        ],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "gilang@gmail.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "rahasia123"
                  },
                  "remember_me": {
                    "type": "boolean",
                    "example": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "token": {
                          "type": "string",
                          "example": "jwt_token"
                        },
                        "user": {
                          "$ref": "#/components/schemas/User"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/GenericError"
          }
        }
      }
    },
...
    "/auth/forgot-password": {
      "post": {
        "summary": "Forgot Password",
        "tags": [
          "Auth"
        ],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "gilang@gmail.com"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        }
      }
    },
    "/auth/reset-password": {
      "post": {
        "summary": "Reset Password",
        "tags": [
          "Auth"
        ],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string",
                    "example": "123456"
                  },
                  "new_password": {
                    "type": "string",
                    "example": "passwordbaru"
                  },
                  "confirm_password": {
                    "type": "string",
                    "example": "passwordbaru"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        }
      }
    },
...
    "/users/me": {
      "get": {
        "summary": "Get My Profile",
        "tags": [
          "Profile"
        ],
        "responses": {
          "200": {
            "description": "Success Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update Profile",
        "description": "Partial update for user profile. PATCH is used because it updates only specific fields without replacing the entire object.",
        "tags": [
          "Profile"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Gilang Baru"
                  },
                  "phone": {
                    "type": "string",
                    "example": "081111111"
                  },
                  "profile_picture": {
                    "type": "string",
                    "example": "profile-1.jpg"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        }
      }
    },
    "/users/me/password": {
      "patch": {
        "summary": "Update Password",
        "tags": [
          "Profile"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "current_password": {
                    "type": "string",
                    "example": "passwordlama"
                  },
                  "new_password": {
                    "type": "string",
                    "example": "passwordbaru"
                  },
                  "confirm_password": {
                    "type": "string",
                    "example": "passwordbaru"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          },
          "400": {
            "$ref": "#/components/responses/GenericError"
          }
        }
      }
    },
...
    "/users/me/request-email-otp": {
      "post": {
        "summary": "Request Email Change OTP",
        "tags": [
          "Profile"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "new_email": {
                    "type": "string",
                    "example": "gilang_baru@gmail.com"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        }
      }
    },
    "/users/me/email": {
      "put": {
        "summary": "Verify OTP & Update Email",
        "tags": [
          "Profile"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "otp": {
                    "type": "string",
                    "example": "123456"
                  },
                  "new_email": {
                    "type": "string",
                    "example": "gilang_baru@gmail.com"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          },
          "400": {
            "$ref": "#/components/responses/GenericError"
          }
        }
      }
    },
    "/users/managers": {
      "post": {
        "summary": "Create Manager (Admin Only)",
        "tags": [
          "Users (Employee Management)"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "branch_id": {
                    "type": "integer",
                    "example": 1
                  },
                  "name": {
                    "type": "string",
                    "example": "Budi Manager"
                  },
                  "email": {
                    "type": "string",
                    "example": "manager@example.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "password123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        }
      }
    },
    "/users/cashiers": {
      "post": {
        "summary": "Create Cashier (Manager Only)",
        "tags": [
          "Users (Employee Management)"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Siti Kasir"
                  },
                  "email": {
                    "type": "string",
                    "example": "kasir@example.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "password123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        }
      }
    },
    "/users/{id}/disable": {
      "patch": {
        "summary": "Disable/Deactivate User (Admin Only)",
        "tags": [
          "Users (Employee Management)"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        }
      }
    },
    "/categories": {
      "get": {
        "summary": "Get All Categories",
        "tags": [
          "Catalogues"
        ],
        "security": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Category"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Category (Admin Only)",
        "tags": [
          "Catalogues"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        }
      }
    },
    "/categories/{id}": {
      "put": {
        "summary": "Update Category (Admin Only)",
        "tags": [
          "Catalogues"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        }
      },
      "delete": {
        "summary": "Delete Category (Admin Only)",
        "tags": [
          "Catalogues"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        }
      }
    },
    "/materials": {
      "get": {
        "summary": "Get All Materials (Admin/Manager Only)",
        "tags": [
          "Catalogues"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Material"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Material (Admin Only)",
        "tags": [
          "Catalogues"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaterialRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        }
      }
    },
    "/materials/{id}": {
      "put": {
        "summary": "Update Material (Admin Only)",
        "tags": [
          "Catalogues"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaterialRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        }
      },
      "delete": {
        "summary": "Delete Material (Admin Only)",
        "tags": [
          "Catalogues"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        }
      }
    },
    "/products": {
      "get": {
        "summary": "Get All Products",
        "description": "Public users see only basic info. Does NOT include BOM/Recipe.",
        "tags": [
          "Catalogues"
        ],
        "security": [],
        "parameters": [
          {
            "name": "category_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Product"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Product with Recipe (Admin Only)",
        "tags": [
          "Catalogues"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "summary": "Get Product Detail",
        "description": "Includes Recipe (BOM) only if requester is Admin or Manager and adds ?include_recipe=true.",
        "tags": [
          "Catalogues"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "include_recipe",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Product"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update Product and Recipe (Admin Only)",
        "tags": [
          "Catalogues"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        }
      },
      "delete": {
        "summary": "Soft Delete Product (Admin Only)",
        "tags": [
          "Catalogues"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        }
      }
    },
    "/branches": {
      "get": {
        "summary": "Get All Branches",
        "description": "Fetch a list of all branches. Public users see only active branches. Admins see all branches.",
        "tags": [
          "Branches"
        ],
        "security": [],
        "parameters": [
          {
            "name": "include_inactive",
            "in": "query",
            "description": "If true, includes inactive branches (Admin only).",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Branch"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Branch (Admin Only)",
        "tags": [
          "Branches"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBranchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        }
      }
    },
    "/branches/{id}": {
      "put": {
        "summary": "Update Branch (Admin Only)",
        "tags": [
          "Branches"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBranchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        }
      },
      "delete": {
        "summary": "Soft Delete Branch (Admin Only)",
        "description": "Sets is_active to false instead of physical deletion.",
        "tags": [
          "Branches"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        }
      }
    },
...
    "/uploads": {
      "post": {
        "summary": "Upload a File",
        "description": "Upload a file to the server. Supports JPG, JPEG, PNG, and WEBP. Maximum size is 5MB.",
        "tags": [
          "Media"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "The file to upload (Max 5MB)"
                  },
                  "folder": {
                    "type": "string",
                    "description": "Optional subfolder name (e.g. 'products', 'profiles')",
                    "default": "misc"
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "file uploaded successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "url": {
                          "type": "string",
                          "example": "http://localhost:8080/uploads/products/1689328293.jpg"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/GenericError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "phone": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "profile_picture": {
            "type": "string"
          }
        }
      },
...
      "Category": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string",
            "example": "Kopi"
          }
        }
      },
      "CategoryRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "Kopi"
          }
        }
      },
      "Material": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "category_id": {
            "type": "integer"
          },
          "name": {
            "type": "string",
            "example": "Biji Kopi Arabica"
          },
          "unit": {
            "type": "string",
            "example": "gram"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "MaterialRequest": {
        "type": "object",
        "required": [
          "category_id",
          "name",
          "unit"
        ],
        "properties": {
          "category_id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Biji Kopi Arabica"
          },
          "unit": {
            "type": "string",
            "example": "gram"
          }
        }
      },
...
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "category_id": {
            "type": "integer"
          },
          "name": {
            "type": "string",
            "example": "Kopi Susu Aren"
          },
          "description": {
            "type": "string",
            "example": "Perpaduan espresso dan aren manis"
          },
          "price": {
            "type": "number",
            "format": "float",
            "example": 20000
          },
          "image_url": {
            "type": "string"
          },
          "is_active": {
            "type": "boolean"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "recipe": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductBOM"
            }
          }
        }
      },
...
    },
    "responses": {
      "GenericError": {
        "description": "Error response",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "GenericSuccess": {
        "description": "Success operation",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Operation success"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}